remove comments and add requied classes

+++ src/main/java/com/kodilla/ecommerce/service/CartServiceImpl.java
@@ -4,20 +4,19 @@ import com.kodilla.ecommerce.domain.Cart;
 import com.kodilla.ecommerce.domain.Product;
 import com.kodilla.ecommerce.exception.NotFoundException;
 import com.kodilla.ecommerce.repository.CartRepository;
+import com.kodilla.ecommerce.repository.OrderRepository;
+import com.kodilla.ecommerce.repository.ProductRepository;
 import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
+@AllArgsConstructor
 public class CartServiceImpl implements CartService {
 
     private final CartRepository cartRepository;
-
-    public CartServiceImpl(CartRepository cartRepository) {
-        this.cartRepository = cartRepository;
-    }
+    private final ProductRepository productRepository;
+    private final OrderRepository orderRepository;
+    private final OrderService orderService;
 
     @Override
     public Cart createCart(Cart cart) {
@@ -36,11 +35,11 @@ public class CartServiceImpl implements CartService {
     @Override
     public void addProductToCart(Long id, Long productId) throws NotFoundException {
         if(cartRepository.existsById(id)) {
-            //odkomentować po dodaniu productRepository
-
-//            cartRepository.findById(id).get().getProducts()
-//                    .add(productRepository.findById(productId)
-//                            .orElseThrow(new NotFoundException("Product id: " + productId + " not found")));
+            if(productRepository.existsById(productId)) {
+                cartRepository.findById(id).get().getProducts().add(productRepository.findById(productId).get());
+            } else {
+                throw new NotFoundException("Product id: " + productId + " not found");
+            }
         } else {
             throw new NotFoundException("Cart id: " + id + " not found");
         }
@@ -49,11 +48,11 @@ public class CartServiceImpl implements CartService {
     @Override
     public void deleteProductFromCart(Long id, Long productId) throws NotFoundException {
         if(cartRepository.existsById(id)) {
-            //odkomentować po dodaniu productRepository
-
-//            cartRepository.findById(id).get().getProducts()
-//                    .remove(productRepository.findById(productId)
-//                    .orElseThrow(new NotFoundException("Product id: " + productId + " not found")));
+            if(productRepository.existsById(productId)) {
+                cartRepository.findById(id).get().getProducts().remove(productRepository.findById(productId).get());
+            } else {
+                throw new NotFoundException("Product id: " + productId + " not found");
+            }
         } else {
             throw new NotFoundException("Cart id: " + id + "not found");
         }
@@ -62,13 +61,12 @@ public class CartServiceImpl implements CartService {
     @Override
     public void createOrderFromCart(Long id) {
         if(cartRepository.existsById(id)) {
-
-            //odkomentować jak już będzie stworzone orderRepository
-
-            //Order order = orderRepository.createOrder(cartRepository.findById(id).get());
-            //if(order.findByCartId(id).isPresent() {
+            orderService.createOrder(cartRepository.findById(id).get());
+            if(orderRepository.findByUserId(cartRepository.findById(id).get().getUser().getId()).isPresent()) {
                 cartRepository.deleteById(id);
-            // }
+             } else {
+                throw new NotFoundException("Create order failed. Order with cart id: " + id + " not found");
+            }
         } else {
             throw new NotFoundException("Cart id: " + id + "not found");
         }
